Abstract away the polling so it can be reused
